requirements.txt
nbformat==5.1.*
docker==5.0.*
google-cloud-aiplatform==1.1.*